+++ src/protect/strenc/layout_types.rs
@@ -10,9 +10,9 @@ pub struct Inputs<'a> {
     pub codeRva: u32,
     pub codeLen: u32,
     pub codeBlob: &'a [u8],
-    pub selfKey: u32,
+    pub selfKey: u64,
     pub lazyInterval: u32,
-    pub masterSeed: u32,
+    pub masterSeed: u64,
 }
 
 pub struct Built {
@@ -36,12 +36,12 @@ pub fn align(content: &mut Vec<u8>, n: usize) {
     }
 }
 
-pub fn patchJunk(content: &mut [u8], seed: u32) {
+pub fn patchJunk(content: &mut [u8], seed: u64) {
     let placeholders = [0xB16B_00B5u32, 0xCAFE_D00D, 0x5EED_1234];
     for (n, ph) in placeholders.iter().enumerate() {
         let pattern = ph.to_le_bytes();
         if let Some(pos) = content.windows(4).position(|w| w == pattern) {
-            let value = super::cipher::deriveKey(seed, 0xABCD_0000 ^ n as u32);
+            let value = super::cipher::deriveKey(seed, 0xABCD_0000 ^ n as u32) as u32;
             content[pos..pos + 4].copy_from_slice(&value.to_le_bytes());
         }
     }
